+++ AlphaEngine/Debris.cpp
@@ -23,6 +23,7 @@ prior written consent of DigiPen Institute of Technology is prohibited.
 #include "WaveManager.h"
 #include <cmath>
 #include "Data.h"
+#include "GameStateList.h"
 
 //VARIABLES
 //#define SPEED_DEBRIS 30
@@ -103,7 +104,7 @@ void Debris::init()
 	OUTERRIM_TO_DEBRIS	= DebrisDataMap["Distance_To_Planet"];
 }
 
-void Debris::update(f64 frame_time)
+void Debris::update(f32 frame_time)
 {
 	for (int j = 0; j < debris_vector_all.size(); j++) {
 		for (size_t i = 0; i < debris_vector_all[j].size(); i++) {
@@ -227,8 +228,6 @@ void Debris::update(f64 frame_time)
 			}
 		}
 	}
-
-
 }
